# api/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the API documentation for the FlashMind application.",
    "version": "1.0.0",
    "title": "FlashMind API"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "tags": [
    {
      "name": "auth",
      "description": "Operations about authentication"
    },
    {
      "name": "flashmind",
      "description": "Operations about flashcards"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Get JWT token",
        "description": "Generates a JWT token for authentication",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "credentials",
            "description": "User credentials",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "example": {
                "username": "admin",
                "password": "password"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token generated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/flashmind": {
      "post": {
        "tags": [
          "flashmind"
        ],
        "summary": "Create a new flashcard",
        "description": "Creates a new flashcard (admin only)",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "card",
            "description": "Flashcard to create",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "question": {
                  "type": "string"
                },
                "answer": {
                  "type": "string"
                },
                "subject": {
                  "type": "string"
                }
              },
              "example": {
                "question": "What is the capital of France?",
                "answer": "Paris",
                "subject": "Geography"
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Flashcard created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Invalid JWT or insufficient permissions"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "flashmind"
        ],
        "summary": "Get all flashcards",
        "description": "Retrieves a list of all flashcards",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "A list of flashcards",
            "schema": {
              "type": "object",
              "properties": {
                "cards": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "question": {
                        "type": "string"
                      },
                      "answer": {
                        "type": "string"
                      },
                      "subject": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Invalid JWT or insufficient permissions"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/flashmind/{id}": {
      "put": {
        "tags": [
          "flashmind"
        ],
        "summary": "Update a flashcard",
        "description": "Updates an existing flashcard (admin only)",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the flashcard to update"
          },
          {
            "in": "body",
            "name": "card",
            "description": "Flashcard data to update",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "question": {
                  "type": "string"
                },
                "answer": {
                  "type": "string"
                },
                "subject": {
                  "type": "string"
                }
              },
              "example": {
                "question": "What is the capital of France?",
                "answer": "Paris",
                "subject": "Geography"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flashcard updated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Flashcard not found"
          },
          "403": {
            "description": "Forbidden - Invalid JWT or insufficient permissions"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "flashmind"
        ],
        "summary": "Delete a flashcard",
        "description": "Deletes a flashcard (admin only)",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the flashcard to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Flashcard deleted successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Flashcard not found"
          },
          "403": {
            "description": "Forbidden - Invalid JWT or insufficient permissions"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  }
}

  